Roll all R2->R3 related model changes into a single migration
