--- db/migrate/20130910153446_ensure_duration_is_correct.rb
@@ -1,22 +1,0 @@
-# Copyright 2011-2013, The Trustees of Indiana University and Northwestern
-#   University.  Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-# 
-# You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-# 
-# Unless required by applicable law or agreed to in writing, software distributed 
-#   under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-#   CONDITIONS OF ANY KIND, either express or implied. See the License for the 
-#   specific language governing permissions and limitations under the License.
-# ---  END LICENSE_HEADER BLOCK  ---
-
-class EnsureDurationIsCorrect < ActiveRecord::Migration
-  def change
-    MediaObject.where({}).each do |media_object|
-      media_object.populate_duration!
-      media_object.save( validate: false )
-    end
-  end
-end
