--- db/migrate/20131212203311_add_dropbox_to_collections.rb
@@ -1,10 +1,0 @@
-class AddDropboxToCollections < ActiveRecord::Migration
-  def up
-    Admin::Collection.find_each({},{batch_size:5}) do |collection|
-      if ! collection.dropbox_directory_name
-        collection.send(:create_dropbox_directory!)
-        collection.save( validate: false )
-      end
-    end
-  end
-end
