+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20140218202613) do
+ActiveRecord::Schema.define(:version => 20140306225510) do
 
   create_table "bookmarks", :force => true do |t|
     t.integer  "user_id",     :null => false
